Import G998BXXU3BUK8 from SM-G998B_12_Opensource.zip

+++ drivers/media/platform/exynos/camera/is-device-csi.h
@@ -141,7 +141,7 @@ struct is_device_csi {
 	struct v4l2_subdev		**subdev;
 	struct phy			*phy;
 
-	u32 error_id[CSI_VIRTUAL_CH_MAX];
+	ulong error_id[CSI_VIRTUAL_CH_MAX];
 	u32 error_id_last[CSI_VIRTUAL_CH_MAX];
 	u32 error_count;
 	u32 error_count_vc_overlap;
